Official-storybook: Add hooks-based state example & embed it in jest
